requirements.txt
alembic
asyncio
asyncpg
fastapi
gunicorn
python-dotenv
SQLAlchemy
uvicorn